// frontend/src/shared/lib/api/model/umapResultDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */
import type { UmapPointDto } from './umapPointDto';
import type { ClusterAssignmentDto } from './clusterAssignmentDto';
import type { RepresentativeDto } from './representativeDto';
import type { UmapResultDtoPickerParams } from './umapResultDtoPickerParams';

export interface UmapResultDto {
  points: UmapPointDto[];
  clusters: ClusterAssignmentDto[];
  representatives: RepresentativeDto[];
  cluster_count: number;
  picker: string;
  picker_params: UmapResultDtoPickerParams;
  skipped_molecule_ids: string[];
}