// frontend/src/shared/lib/api/model/umapJobDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */
import type { UmapJobDtoPickerParams } from './umapJobDtoPickerParams';
import type { UmapJobDtoErrorMessage } from './umapJobDtoErrorMessage';

export interface UmapJobDto {
  id: string;
  status: string;
  picker: string;
  picker_params: UmapJobDtoPickerParams;
  error_message?: UmapJobDtoErrorMessage;
}